Fix quick launcher failing to load when running menu from a game
The game could be a dump, cart or PSN, but using sceClibPrintf fails to write properly to the uri buffer.

This may potienally fix loading in some games since sceClibSnprintf was used in RecursiveMakeDir (which creates the VSH directory so that config files can be stored. Hence without the dir some games may have failed to bring up the menu.)